services/auth/requirements.txt
fastapi[standard]
pydantic
typing
requests
urllib3
python-jose[cryptography]
bcrypt
fastapi-decorators

google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

redis[hiredis]
psycopg[binary]
psycopg_pool

isort[requirements_deprecated_finder]